docker_requirements.txt
aiohttp
aiosignal==1.3.1
# Exclude pywin32 in the Docker environment
# pywin32
Django>=3.2
django-filter

#django-crispy-forms[bootstrap5]==2.0
django-crispy-forms>=1.11,<1.12
crispy_bootstrap5













